Market_Flow/requirements.txt
requests
python-dotenv
pandas
faker
flask
flask-cors
plotly
openpyxl
xlsxwriter
